With r181803 on 2008-08-17 23:27:27Z the first VIMAGE commit went into HEAD.  Enable VIMAGE in GENERIC kernels and some others (where GENERIC does not exist) on HEAD.
Disable building LINT-VIMAGE with VIMAGE being default.

This should give it a lot more exposure in the run-up to 12 to help
us evaluate whether to keep it on by default or not.
We are also hoping to get better performance testing.
The feature can be disabled using nooptions.

Requested by:		many
Reviewed by:		kristof, emaste, hiren
X-MFC after:		never
Relnotes:		yes
Differential Revision:	https://reviews.freebsd.org/D12639


git-svn-id: https://svn.freebsd.org/base/head@324810 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f