Create a shell script to build sys/boot on all the architectures.
One could run this from any directory, but it's designed to do
regression testing on sys/boot (it only tests on a subset of
architectures since all of them would take a lot longer and not help).
This will also ensure that future commits to sys/boot compile
everywhere.

Sponsored by: Netflix


git-svn-id: https://svn.freebsd.org/base/head@324882 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f